feat: 게시글 CRUD 완성

+++ src/main/java/com/sparta/maeng9newsfeed/service/BoardService.java
@@ -1,12 +1,19 @@
 package com.sparta.maeng9newsfeed.service;
 
+import com.sparta.maeng9newsfeed.annotation.Auth;
+import com.sparta.maeng9newsfeed.dto.AuthUser;
 import com.sparta.maeng9newsfeed.dto.BoardResponse;
 import com.sparta.maeng9newsfeed.entity.Board;
 import com.sparta.maeng9newsfeed.entity.BoardLike;
 import com.sparta.maeng9newsfeed.entity.Image;
+import com.sparta.maeng9newsfeed.entity.User;
 import com.sparta.maeng9newsfeed.repository.BoardLikeRepository;
 import com.sparta.maeng9newsfeed.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -24,8 +31,8 @@ public class BoardService {
     private final UserService userService;
 
     @Transactional
-    public BoardResponse create(String content, List<MultipartFile> images) {
-        Board board = new Board(content);
+    public BoardResponse create(String content, List<MultipartFile> images, long userId) {
+        Board board = new Board(content, userService.findByUserId(userId));
 
         Board saveBoard = boardRepository.save(board);
 
@@ -40,25 +47,45 @@ public class BoardService {
         return new BoardResponse(board);
     }
 
+    public Page<BoardResponse> getMyBoardList(long userId) {
+        Pageable pageable = PageRequest.of(0, 10, Sort.by("modifiedAt").descending());
+
+        Page<Board> boardList = boardRepository.findAllByUser_Id(userId, pageable);
+
+        return boardList.map(BoardResponse::new);
+    }
+
     @Transactional
-    public String delete(long boardId) {
-        boardRepository.delete(findBoardById(boardId));
+    public String delete(long boardId, long userId) {
+        Board board = boardRepository.findById(boardId).orElseThrow(
+                () -> new IllegalArgumentException("존재하지 않는 게시물 입니다.")
+        );
+        if (board.getUser().getId() == userId) {
+            boardRepository.delete(findBoardById(boardId));
+        } else {
+            throw new IllegalArgumentException("게시글을 작성한 사용자가 아닙니다.");
+        }
+
         return "게시글 삭제 완료";
     }
 
-    private Board findBoardById(long boardId) {
+    public Board findBoardById(long boardId) {
         return boardRepository.findById(boardId).orElseThrow(
                 () -> new IllegalArgumentException("게시글을 찾을 수 없습니다.")
         );
     }
 
     @Transactional
-    public long update(long boardId, String content, List<MultipartFile> images) {
+    public long update(long boardId, String content, List<MultipartFile> images, long userId) {
         Board board = boardRepository.findById(boardId).orElseThrow(
                 () -> new IllegalArgumentException("존재하지 않는 게시물 입니다.")
         );
-        board.updateBoard(content);
-        imageService.updateImage(board, images);
+        if (board.getUser().getId() == userId) {
+            board.updateBoard(content);
+            imageService.updateImage(board, images);
+        } else {
+            throw new IllegalArgumentException("게시글을 작성한 사용자가 아닙니다.");
+        }
 
         return boardId;
     }
